Ethernet: Ethernet feature added
This patch is enabled Ethernet network support with DHCP
or static IP configuration.

It is does not matter which way ethernet is plugged to the board.
It could be plugged to internal Ethernet chip or through
the USB/Ethernet adapter.

android.hardware.ethernet permission need to be presented on the
board to be able use ethernet feature.

Change-Id: Ib294c3ef755510f250247ff37768f52d8a2f8b04
Signed-off-by: Anatolii Shuba <anatolii.shuba@ti.com>